fix: move buildConnectionString to separate file for client import
Templates module uses node:fs which can't be imported in browser.
